requirements.txt
fastapi[all]
pydantic
sqlalchemy
psycopg2-binary
alembic
python-dotenv
pytest